requirements/test-ci.txt
pytest-cov
codecov
redis
PyYAML
msgpack-python>0.2.0
